requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
mysql-connector==2.2.9
mysqlclient==2.2.6
mysqlx-connector-python==9.1.0
packaging==24.2
protobuf==4.25.3
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
timedelta==2020.12.3
typing_extensions==4.12.2
Werkzeug==3.1.3
